Go back to use `enum` keyword for enums. (#113)
* Go back to use `enum` keyword for enums.

Previously we used an enum-like pattern so that we could attach
methods like toJSON/fromJSON to the enum itself, i.e.
StateEnum.fromJSON.

This was done out of a) convenience b/c we already knew where the
`StateEnum` came from (i.e. had it imported) and b) in-theory
developer ergonomics for calling the `StateEnum.fromJSON` method.

However, a) is easy to fix with just trying harder, i.e. a new
getEnumMethod that can find top-level functions for a given enum,
i.e. stateEnumFromJSON is in module x/y/z.

And b) is not really a factor b/c developers should rarely serialize
an enum directly and are probably/basically always working with
MessageFoo.to/fromJSON, and it's only the internals of
MessageFoo.to/fromJSON that have to know where the enum functions
live/how to call them.

So, this PR negates a) and ignores b), especially because developers
really do expect the protobuf enums to TS enums, i.e. for doing
things like Object.entries(StateEnum).

* Format.