refactor: Switch to @grpc/grpc-js altogether. Bump to nestjs v8. (#401)
* Remove grpc in integrations as well. Bump to nestjs v8.

grpc is deprecated and the authors recommend to use @grpc/grpc-js instead.
While ts-proto does that almost everywhere, it still needed grpc for the
nestjs integration tests.

Since nestjs switched to v8, it uses @grpc/grpc-js as well, so bump to
nestjs v8. This also requires rxjs v7+.

* Change port number for nestjs-simple-restparameters test.

For me, it occassionally failed when running it together with the nestjs-simple
test, perhaps because the two shared a port number -- they never failed when running
them simultaneously. Perhaps there is some changed behaviour between nestjs 7 and 8?

* Fix test code on node 12.