feat(options): adds protoc-gen-go-like M option
This commit adds a new feature that aims to provide the same
functionality as `M` for the go compiler (https://developers.google.com/protocol-buffers/docs/reference/go-generated#package).
I opted for not modifying the type map, because we'll also want to
figure out whether or not we need to generate the file, so we'll be
accessing the `options.M` everywhere and I found it less confusing to
always get mapping information from `options.M` instead of `options.M`
or the type map depending on the decision that we're making (we need to
know if we need to do a relative import or just the raw provided
import too).

Fixes #596