feat(client): allow overriding the service identifier
Currently the service identifier is the fully qualified "fullName" of the
service, for example, "simple.Echoer"

This identifier is hard-coded in the auto-generated client with string literals.

Instead, allow overriding the service ID with a new optional argument in the
client constructor, default the service ID to the fullName, and store the ID in
a read-only field on the client object.

This allows the user to override the service ID if desired while still
maintaining the old behavior otherwise.

Signed-off-by: Christian Stewart <christian@paral.in>