fix: export client RPC interfaces
Add client RPC interfaces to codegen exports.

Client implementations accept an `Rpc` that varies. The type is an
especially useful export for consumers that wish to provide their own
`Rpc` implementation. This patch simply adds the `export` keyword to
those interfaces.

This patch also updates all the test snapshots, some of which appear to
have shifted independent of this patch. I'm happy to drop those if
unwanted.