Perf: oneof union encode to switch statement
Previously each oneof union case generated its own `if` statement in the
encode method which becomes very inefficient with more cases as each of
the conditions needed to be evaluated but only one would ever be true.
They were not even `else if (...)` so once a match was found the rest of
the `if` statements would still continue to be evaluated.