feat: allow `$type` to be optional (#1013)
### Description

Add another possible value to the `outputTypeAnnotations` option:
`optional`. This makes type definitions on interfaces optional, which
may be useful if you want to use the `$type` field for runtime type
checking on responses from a server but don't want to have to set the
$type on each request message you create.

### Testing

Ensured this causes no regressions for existing integration tests. Also
added new integration tests to prevent future regressions.