fix: `outputSchema=true` when `onlyTypes=true` (#1039)
This PR aims to fix a conflict that is created between two options that
results in code that fails to compile. Please note that this is not an
attempt to fix the underlying problem, which I have explained in a TODO
comment in the code for posterity. I also created new tests to capture
any regressions.

---------

Co-authored-by: semantic-release-bot <semantic-release-bot@martynus.net>