fix: camelToSnake to respect uppercase words, such as "GetAPIValue" -> "GET_API_VALUE" (#1046)
## Problem
`camelToSnake()` is converting the example string "TestAPI" ->
"TEST_AP_I" when converting a 3rd party integration proto files that I
am attempting to generate with the Nest option. I believe that the
conversion should respect the "API" part of the string and not split it
into two parts like so.

## Fix
I swapped the logic to use the already imported "case-anything"
library's `snakeCase()` function as `snakeCaseAnything()` and added a
couple of simple tests to ensure the logic.

I would love for this fix to go through as it is a pretty inconvenient
problem at the moment (:

Thanks!