chore: Update Jest in integration/simple-esmodule-interop (#1054)
The test `integration/simple-esmodule-interop` runs with its own
versions of Jest and TypeScript.

But the Jest version is pretty old (from 2020), and does not support the
package.json "exports" field correctly. With this version of Jest, it's
not possible to import from packages that use the exports field for dual
packages.

This updates to a more recent version (^29.6.1, same as root
package.json) to support the exports field. (28 added support, see the
[release
notes](https://jestjs.io/blog/2022/04/25/jest-28#packagejson-exports)).