fix: Add check for lower bound with forceLong=number (#1057)
With default plugin options, or `useJsTypeOverride` for a field with
option `jstype = JS_NUMBER`, 64-bit integral values are converted to a
JavaScript Number when decoding from binary. When a value is too large
for Number, an error is thrown.

But signed integers can also be too _small_ for Number. This PR adds a
check for the lower bound.