feat: add option `noDefaultsForOptionals` (#1051)
Adds a new option - `noDefaultsForOptionals` that allows for
compatibility with [Square/Block's Wire Protobuf
implementation](https://square.github.io/wire/) with respect to
defaulting of optional fields.

Note: As per the documentation added to the readme, this option provides
non-standard protobuf behaviour so it would be understandable if this
isn't something you'd like in the library. I'm hoping with the
appropriate warning in the docs that you'll consider this as as option.