Remove longToBigint
BinaryReader already returns a BigInt for 64-bit integers. We don't need to convert it.

If BigInt is not supported in the current runtime (Safari was the last to catch up in 2020), it returns a string. In that case, it isn't possible to convert to a BigInt anyway.

This should be a regression, because the old code relied on BigInt as well.