feat: bigIntLiteral option for using BigInt literals (#1063)
Fixes #928. Since I finally got around to contributing, I figured I'd
try this old issue out too, and it was surprisingly simple. Entirely up
to you whether to trash it or not - it's not a very consequential option
and the size optimization should probably be handled by
https://github.com/terser/terser/issues/1535 anyway, and it could cause
some breakage for users.

Supercedes #932.