perf: Generating "else if" where applicable (#1141)
In oneof scenarios, this commit proposes generating "else if" statements
where applicable.

The original idea was an attempt to speed up Typescript compilation by
reducing the amount of paths Typescript can analyze. Spoiler alert =>
nothing changes from that point of view.

The compilation of my [proto
file](https://github.com/workadventure/workadventure/blob/develop/messages/protos/messages.proto)
takes between 41 and 44 seconds with and without "else if" statements.

Still, the "else if" conditions should slightly speed up the runtime
execution, by avoiding unnecessary checks to the Javascript runtime.

This PR is directly linked to
https://github.com/stephenh/ts-proto/issues/1135
Note: in https://github.com/stephenh/ts-proto/issues/1135, @stephenh you
recommend trying to use a switch statement (instead of "else if"). This
would require a really big refactoring and I'm not familiar enough with
the code base to attempt that.
The "else if" implementation was a bit of a low hanging fruit.

I still think the "//ts-nocheck" annotation would be really useful and
is the way to go, since it would simply take "0 seconds" to typecheck
the generated file (you can't beat 0 in terms of performance :laughing:
)