perf: Replacing "else if" with a "switch case" statement to improve Typescript performance (#1142)
This is a follow-up (and solution) to #1135 and #1141

I generated a trace from the build process and using
[typescript-analyze-trace](https://github.com/microsoft/typescript-analyze-trace)
to analyze the hotspots.

It turns out that time is spent almost exclusively in this kind of code:

```
if (object.choice?.$case === "aNumber" && object.choice?.value !== undefined && object.choice?.value !== null) {
  message.choice = { $case: "aNumber", value: object.choice.value };
} else if (
  object.choice?.$case === "aString" && object.choice?.value !== undefined && object.choice?.value !== null
) {
  message.choice = { $case: "aString", value: object.choice.value };
} else if (
//...
```

The previous PR I opened adding "else if" does nothing regarding
Typescript perf.

With "else if", my test file trace looks like this:

```
npx analyze-trace traceDir
Hot Spots
├─ Check file /home/dan/projects/ts-proto/integration/large/messages.ts (31108ms)
│  ├─ Check deferred node from (line 6267, char 3) to (line 6539, char 4) (13123ms)
│  ├─ Check deferred node from (line 15342, char 3) to (line 15591, char 4) (7092ms)
│  ├─ Check deferred node from (line 10849, char 3) to (line 11059, char 4) (4460ms)
│  ├─ Check deferred node from (line 17320, char 3) to (line 17496, char 4) (2146ms)
│  ├─ Check deferred node from (line 3720, char 3) to (line 3832, char 4) (610ms)
│  └─ Check deferred node from (line 8791, char 3) to (line 8915, char 4) (516ms)
└─ Check file /home/dan/projects/ts-proto/node_modules/typescript/lib/lib.dom.d.ts (1109ms)
```

Now, with this commit, I'm replace `else if` with a `switch` statement.

The analysis of the build looks like this:

```
Hot Spots
├─ Check file /home/dan/projects/ts-proto/integration/large/messages.ts (3139ms)
│  └─ Check deferred node from (line 6278, char 3) to (line 6573, char 4) (525ms)
└─ Check file /home/dan/projects/ts-proto/node_modules/typescript/lib/lib.dom.d.ts (617ms)
```

The biggest `switch` statement takes 525ms to analyze (down from
13seconds with `else if`)

Implementation-wise, the hardest part is knowing when to close the
switch statement. I'm doing a test at the beginning of each loop and at
the very end of the function.