feat: add option for using Temporal timestamp representation (#1219)
This PR adds support for a new `useDate` option: `temporal`. When used,
it will cause the generated code to use the new [Temporal.Instant
](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Temporal/Instant)
object.

The advantages of supporting the new API include higher precision
timestamps (up to nanosecond precision, which aligns with the Protobuf
spec), as well as a variety of Temporal-related utilities, such as
[Duration](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Temporal/Duration).
In a future update, it may be useful to add a similar flag for using the
`Temporal.Duration` object for `Duration` representation.