fix: fix double division, fix loss of millisecond unit (#1221)
This PR fixes an issue I introduced in the initial `useDate=temporal` PR
[here](https://github.com/stephenh/ts-proto/pull/1219). The generated
code for `toTimestamp` was dividing the `epochMilliseconds` twice,
causing the date to get truncated to `1970-01-01`, so the tests would
seem to work for dates.

There was _also_ a bug in how `nanos` was calculated:
```typescript
Temporal.Instant
    .fromEpochMilliseconds(instant.epochMilliseconds)
    .until(instant);
```
since we were deriving the start date from `epochMilliseconds`, so the
remainder only included micro/nanoseconds (milliseconds were always
`000`).

The PR addresses the first issue by ensuring we only divide once, in the
generated `seconds` code. It fixes the latter by using the built-in
[Temporal.Instant.prototype.round](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Temporal/Instant/round)
method to round to the nearest second without needing to do any explicit
math.

Lastly, it also updates the tests, as the `toStrictEqual` check was
succeeding even for `Instant`s whose string representations were clearly
different, so now we will test against the output string.